move partitioning to own file; rest of joint_infer.jl into ParallelRun.jl
